refactor: Refactored all routes to reduce codebase
