--- src/app/api/media/schemas.py
@@ -1,17 +1,0 @@
-from datetime import datetime
-from pydantic import BaseModel, Field, validator
-from app.db import MediaType
-
-
-class MediaIn(BaseModel):
-    device_id: int = Field(..., gt=0)
-    type: MediaType = MediaType.image
-
-
-class MediaOut(MediaIn):
-    id: int = Field(..., gt=0)
-    created_at: datetime = None
-
-    @validator('created_at', pre=True, always=True)
-    def default_ts_created(cls, v):
-        return v or datetime.utcnow()
