+++ src/app/api/routes/ping.py
