--- src/app/api/users/__init__.py
@@ -1,1 +1,0 @@
-from .routing import router
